// web/gen/api.ts
export interface Comment {
    /**
     * string comment_id = 1    [json_name = "commentId"];
     *
     * @generated from protobuf field: string account_id = 1
     */
    accountId: string;
    /**
     * @generated from protobuf field: string content = 2
     */
    content: string;
    /**
     * @generated from protobuf field: string sig = 3
     */
    sig: string;
    /**
     * @generated from protobuf field: string public_key = 4
     */
    publicKey: string;
    /**
     * @generated from protobuf field: uint32 key_type = 5
     */
    keyType: number;
    /**
     * @generated from protobuf field: string created_at = 6
     */
    createdAt: string;
}
// ...
export interface CreateCommentRequest {
    /**
     * @generated from protobuf field: string market_id = 1
     */
    marketId: string;
    /**
     * @generated from protobuf field: string account_id = 2
     */
    accountId: string;
    /**
     * @generated from protobuf field: string content = 3
     */
    content: string;
    /**
     * @generated from protobuf field: string sig = 4
     */
    sig: string;
    /**
     * @generated from protobuf field: string public_key = 5
     */
    publicKey: string;
    /**
     * @generated from protobuf field: uint32 key_type = 6
     */
    keyType: number;
}
// ...
class Comment$Type extends MessageType<Comment> {
    constructor() {
        super("api.Comment", [
            { no: 1, name: "account_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "content", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "sig", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 4, name: "public_key", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 5, name: "key_type", kind: "scalar", T: 13 /*ScalarType.UINT32*/ },
            { no: 6, name: "created_at", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<Comment>): Comment {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.accountId = "";
        message.content = "";
        message.sig = "";
        message.publicKey = "";
        message.keyType = 0;
        message.createdAt = "";
        if (value !== undefined)
            reflectionMergePartial<Comment>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Comment): Comment {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string account_id */ 1:
                    message.accountId = reader.string();
                    break;
                case /* string content */ 2:
                    message.content = reader.string();
                    break;
                case /* string sig */ 3:
                    message.sig = reader.string();
                    break;
                case /* string public_key */ 4:
                    message.publicKey = reader.string();
                    break;
                case /* uint32 key_type */ 5:
                    message.keyType = reader.uint32();
                    break;
                case /* string created_at */ 6:
                    message.createdAt = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Comment, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string account_id = 1; */
        if (message.accountId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.accountId);
        /* string content = 2; */
        if (message.content !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.content);
        /* string sig = 3; */
        if (message.sig !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.sig);
        /* string public_key = 4; */
        if (message.publicKey !== "")
            writer.tag(4, WireType.LengthDelimited).string(message.publicKey);
        /* uint32 key_type = 5; */
        if (message.keyType !== 0)
            writer.tag(5, WireType.Varint).uint32(message.keyType);
        /* string created_at = 6; */
        if (message.createdAt !== "")
            writer.tag(6, WireType.LengthDelimited).string(message.createdAt);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message api.Comment
 */
export const Comment = new Comment$Type();
// ...
class CreateCommentRequest$Type extends MessageType<CreateCommentRequest> {
    constructor() {
        super("api.CreateCommentRequest", [
            { no: 1, name: "market_id", kind: "scalar", T: 9 /*ScalarType.STRING*/, options: { "validate.rules": { string: { pattern: "(?i)^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$" } } } },
            { no: 2, name: "account_id", kind: "scalar", T: 9 /*ScalarType.STRING*/, options: { "validate.rules": { string: { pattern: "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)$" } } } },
            { no: 3, name: "content", kind: "scalar", T: 9 /*ScalarType.STRING*/, options: { "validate.rules": { string: { minLen: "1", maxLen: "1000" } } } },
            { no: 4, name: "sig", kind: "scalar", T: 9 /*ScalarType.STRING*/, options: { "validate.rules": { string: { pattern: "^[A-Za-z0-9+/]{20,100}={0,2}$" } } } },
            { no: 5, name: "public_key", kind: "scalar", T: 9 /*ScalarType.STRING*/, options: { "validate.rules": { string: { pattern: "^(04[0-9a-fA-F]{128}|0[23][0-9a-fA-F]{64}|[0-9a-fA-F]{64})$" } } } },
            { no: 6, name: "key_type", kind: "scalar", T: 13 /*ScalarType.UINT32*/, options: { "validate.rules": { uint32: { in: [1, 2] } } } }
        ]);
    }
    create(value?: PartialMessage<CreateCommentRequest>): CreateCommentRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.marketId = "";
        message.accountId = "";
        message.content = "";
        message.sig = "";
        message.publicKey = "";
        message.keyType = 0;
        if (value !== undefined)
            reflectionMergePartial<CreateCommentRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: CreateCommentRequest): CreateCommentRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string market_id */ 1:
                    message.marketId = reader.string();
                    break;
                case /* string account_id */ 2:
                    message.accountId = reader.string();
                    break;
                case /* string content */ 3:
                    message.content = reader.string();
                    break;
                case /* string sig */ 4:
                    message.sig = reader.string();
                    break;
                case /* string public_key */ 5:
                    message.publicKey = reader.string();
                    break;
                case /* uint32 key_type */ 6:
                    message.keyType = reader.uint32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: CreateCommentRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string market_id = 1; */
        if (message.marketId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.marketId);
        /* string account_id = 2; */
        if (message.accountId !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.accountId);
        /* string content = 3; */
        if (message.content !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.content);
        /* string sig = 4; */
        if (message.sig !== "")
            writer.tag(4, WireType.LengthDelimited).string(message.sig);
        /* string public_key = 5; */
        if (message.publicKey !== "")
            writer.tag(5, WireType.LengthDelimited).string(message.publicKey);
        /* uint32 key_type = 6; */
        if (message.keyType !== 0)
            writer.tag(6, WireType.Varint).uint32(message.keyType);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message api.CreateCommentRequest
 */
export const CreateCommentRequest = new CreateCommentRequest$Type();